+++ home/import_helpers.py
@@ -44,6 +44,20 @@ class ImportException(Exception):
         super().__init__()
 
 
+class ImportWarning:
+    """
+    Base warnings for all import related issues.
+    """
+
+    def __init__(
+        self,
+        message: str | list[str],
+        row_num: int | None = None,
+    ):
+        self.row_num = row_num
+        self.message = message
+
+
 def wagtail_to_formdata(val: Any) -> Any:
     """
     Convert a model dict field that may be a nested streamfield (or associated