+++ home/views.py
@@ -180,10 +180,12 @@ class ContentUploadThread(UploadThread):
         super().__init__(**kwargs)
 
     def run(self):
+        warnings = []
         try:
-            import_content(
+            importer = import_content(
                 self.file, self.file_type, self.progress_queue, self.purge, self.locale
             )
+
         except ImportException as e:
             self.result_queue.put(
                 (
@@ -198,7 +200,23 @@ class ContentUploadThread(UploadThread):
             self.result_queue.put((messages.ERROR, ["Content import failed"]))
             logger.exception("Content import failed")
         else:
-            self.result_queue.put((messages.SUCCESS, ["Content import successful"]))
+            warnings = importer.import_warnings
+            if warnings:
+                self.result_queue.put(
+                    (
+                        messages.WARNING,
+                        [
+                            "Content import successful",
+                            *[
+                                f"row {warning.row_num}: {warning.message}"
+                                for warning in warnings
+                            ],
+                        ],
+                    )
+                )
+            else:
+                self.result_queue.put((messages.SUCCESS, ["Content import successful"]))
+
         # Wait until the user has fetched the result message to close the thread
         self.result_queue.join()